Treat wbindgen_describe_closure as a marker
We're already looking at all functions that call `__wbindgen_describe_closure` as special ones that need to be modified.

Instead of passing the `describe::<T>` as a callback, make the function that calls `__wbindgen_describe_closure` a descriptor itself.

Removing the indirection allows to remove a whole bunch of cleanup code, as well as allows to replace the entire function with an import instead of having another layer of call indirection.

It also gives us more flexibility with the signature of said marker later on.