Separate closure handling and cast handling
Closure handling currently conflates two different tasks:

1. Describing closures to the binding system when they're passed to `fn(&Closure<dyn Fn(SomeType)>)`.
2. "Erasing" generic closure types and converting them to a generic JS value (externref) for `Closure::into_js_function`.

There is nothing special about (1) - we already have handling for `&dyn Fn(SomeType)` which is almost identical in every regard to what we need, except for ownership handling (freeing the state object at the end of the call vs registering it in the FinalizationRegistry).

That these two types got completely different handling paths in wasm-bindgen with almost identical implementations is likely a historical artifact rather than intentional choice.

The (2) is more interesting - we don't have anything elsewhere in the wasm-bindgen ecosystem, and if you squint at it hard enough, you'll see that it doesn't care about closures at all, and it's more of a generic "some type to JS externref" cast instruction. That one does need special handling because Rust type system is currently too restrictive and doesn't allow e.g. declaring generic externs. So we need to communicate the materialised descriptors to postprocessing in some other way, which is what the system above was doing.

In this commit I'm starting the work on separating these two concepts, and making the `Closure` type "less special", so its `WasmDescribe` implementation describes the concrete type, and the casting to JS externref only takes the whole `&Closure` as a generic parameter to a cast intrinsic, which doesn't care anymore whether it works on closures or any other type.

In the next step I'm going lift it up even further to a proper generic cast intrinsic.