Erase type of __wbindgen_describe_closure
Since it's just a marker and we use the parent function as source of truth for the import type, we don't need to have Closure arguments here as well - they were unused in the interpreter anyway.

While at it, we can also erase the return type by marking it as never type - this way it can be compatible with any cast targets in the future.