Fix ABI and optimiser handling of `breaks_if_inlined`
This took a horribly long time to debug, but TL;DR is that even with `inline(never)` and `extern "C"` the compiler can change shape of the function if it thinks it has seen all the usages, which it had right below.

This can lead to surprising assertion failures in wasm-bindgen transforms where cast function is suddenly `fn() -> ()` even though descriptor says otherwise.

Passing the pointer to `inform` instead of the previous `null`/`0` placeholder seems enough to convince it to keep the function untouched.