Implement arbitrary generic casts (#4602)
* Treat wbindgen_describe_closure as a marker

We're already looking at all functions that call `__wbindgen_describe_closure` as special ones that need to be modified.

Instead of passing the `describe::<T>` as a callback, make the function that calls `__wbindgen_describe_closure` a descriptor itself.

Removing the indirection allows to remove a whole bunch of cleanup code, as well as allows to replace the entire function with an import instead of having another layer of call indirection.

It also gives us more flexibility with the signature of said marker later on.

* Separate closure handling and cast handling

Closure handling currently conflates two different tasks:

1. Describing closures to the binding system when they're passed to `fn(&Closure<dyn Fn(SomeType)>)`.
2. "Erasing" generic closure types and converting them to a generic JS value (externref) for `Closure::into_js_function`.

There is nothing special about (1) - we already have handling for `&dyn Fn(SomeType)` which is almost identical in every regard to what we need, except for ownership handling (freeing the state object at the end of the call vs registering it in the FinalizationRegistry).

That these two types got completely different handling paths in wasm-bindgen with almost identical implementations is likely a historical artifact rather than intentional choice.

The (2) is more interesting - we don't have anything elsewhere in the wasm-bindgen ecosystem, and if you squint at it hard enough, you'll see that it doesn't care about closures at all, and it's more of a generic "some type to JS externref" cast instruction. That one does need special handling because Rust type system is currently too restrictive and doesn't allow e.g. declaring generic externs. So we need to communicate the materialised descriptors to postprocessing in some other way, which is what the system above was doing.

In this commit I'm starting the work on separating these two concepts, and making the `Closure` type "less special", so its `WasmDescribe` implementation describes the concrete type, and the casting to JS externref only takes the whole `&Closure` as a generic parameter to a cast intrinsic, which doesn't care anymore whether it works on closures or any other type.

In the next step I'm going lift it up even further to a proper generic cast intrinsic.

* Erase type of __wbindgen_describe_closure

Since it's just a marker and we use the parent function as source of truth for the import type, we don't need to have Closure arguments here as well - they were unused in the interpreter anyway.

While at it, we can also erase the return type by marking it as never type - this way it can be compatible with any cast targets in the future.

* Update name to describe_cast and update comments

* Extract casting as rt::wbg_cast

* Remove unnecessary Option<> from interpret_descriptor

* Remove obsolete assertion

Closure descriptors can have arbitrary signatures.

* Support converting to arbitrary values in wbg_cast

* Migrate away from wbg_cast! macro

* Add back __wbindgen_object_clone_ref

I removed it in #4579 without realising that it's an actual intrinsic inserted by externref-xform that can be invoked without going via JS (on modern engines), so it's more efficient to keep it.

* Fix ABI and optimiser handling of `breaks_if_inlined`

This took a horribly long time to debug, but TL;DR is that even with `inline(never)` and `extern "C"` the compiler can change shape of the function if it thinks it has seen all the usages, which it had right below.

This can lead to surprising assertion failures in wasm-bindgen transforms where cast function is suddenly `fn() -> ()` even though descriptor says otherwise.

Passing the pointer to `inform` instead of the previous `null`/`0` placeholder seems enough to convince it to keep the function untouched.

* Use hash for wbindgen_cast suffix for stability

* Add debug comments to cast intrinsics

* Remove visit_return_call handler

Our `wbindgen_describe_cast` is never in the tail position.

* Remap function ids in table elements too

* Fix function replacement

Reuse the import deduplication mechanism instead of doing our own thing.

* Remove VectorIntoJsValue

It can now be entirely covered by the generic cast.