Inline `__externref_heap_live_count`
Instead of declaring it as a C export in the externref runtime module, having the externref transform search for it and the "proper" `__wbindgen_externref_heap_live_count` intrinsic, and having the transform replace one with another, we can just use a `cfg(target_feature = "reference-types")` to choose the right implementation in the first place.

This allows to remove quite a bit of no longer used transitive code for such a small function.