Move Closure's ownership entirely to JS
This makes its memory management a bit less weird and more in line with normal Rust - there is a single side (JS) owning the actual heap data, and it's the only one responsible for invoking the destructor when refcount reaches zero, doesn't matter how that happens.

That allows to remove another `unsafe` too.