Deterministic output
This change addresses non-deterministic ordering in the generated output that
could vary based on compiler settings (eg: number of codegen units is the one
I'm most familiar with) by sorting all the things!

The original use case of deterministic output was to facilitate snapshot
testing of the generated typescript types in downstream projects. Discussion in #4738
broaden the scope to include all outputs, not just types.