Deterministic output
The original use case of deterministic output was to facilitate snapshot
testing of the generated typescript types in downstream projects. Discussion in #4738
broaden the scope to include all outputs, not just types.

Note even with this PR the underlying Wasm can still differ between builds with
different codegen-units settings because how the compiler optimizes based on
codegen-units will impact the Wasm output.

However, the interface (exports, imports, JS bindings, TypeScript definitions)
is now fully deterministic.

This determinism is accomplished a few ways:

- hash suffixes (eg: `__wbindgen_closure__destroy__h7a3b9c2d1e4f5678`) are
  normalized with a sequential counter (like how what was previously done in
  test code): `__wbindgen_closure__destroy__h0000000000000000`
- Wasm exports are sorted by name and function signature
- JS imports and exports are sorted

The normalization relies on walrus's `Type` having a consistent ordering. So if
we have an opinionated sort order, we can define our own, otherwise we can keep
walrus's.