Add feature to wrap JS->Rust calls in catch_unwind
We introduce a new catch-unwind feature, and a new `PanicError` class. When built with
unwinding and std support, we catch panics and raise into JS as `PanicError`s. If the
entrypoint is async, we reject the promise with the `PanicError`.

This can only work if built with `panic=unwind`, which also requires `-Zbuild-std`, so
it only works on a nightly compiler.

The main limitation of this is that it implies that all arguments to functions must be
unwind safe.