Perf: remove redundant demangling string formatting
I ran wasm-bindgen through some input where it took a couple seconds to process
and noticed demangle had some extra `to_string()` calls, so this commit reduces
the number of times we're calling to_string() on a Demangle to 1.

Results of `demangle` as a portion of total sampled time

Before: 12%
After: 7.8%

The single display impl for rustc_demangle::Demangle now dominates, but it's a
small improvement!